log for 2 3
