+++ fairpy/course_allocation/algorithm2.py
@@ -4,7 +4,27 @@ from Course import Course
 from Student import Student
 from functools import cmp_to_key
 import doctest
+import logging
 
+logger = logging.getLogger(__name__)
+console = logging.StreamHandler()  # writes to stderr (= cerr)
+logfile = logging.FileHandler("my_logger2.log", mode="w") 
+logger.handlers = [console,logfile]
+logfile.setFormatter(logging.Formatter('%(asctime)s: %(levelname)s: %(name)s: Line %(lineno)d: %(message)s'))
+
+logger.setLevel(logging.DEBUG)
+console.setLevel(logging.WARNING)
+# console.setLevel(logging.WARNING)
+# logfile = logging.FileHandler("./my_logger3.log", mode="w") 
+
+
+# logger = logging.getLogger(__name__)
+# console = logging.StreamHandler()  # writes to stderr (= cerr)
+# # logfile = logging.FileHandler("my_logger2.log", mode="w") 
+# logger.handlers = [console]
+# # logfile.setFormatter(logging.Formatter('%(asctime)s: %(levelname)s: %(name)s: Line %(lineno)d: %(message)s'))
+# console.setLevel(logging.WARNING)
+# logfile.setLevel(logging.DEBUG)
 
 def csp_mapping(students:list[Student],courses:list[Course]):
     '''
@@ -79,6 +99,12 @@ def algorithm2(price_vector:list[float], maximum:int, eps:float, csp_mapping:cal
         for pre in s.preferences:
             a.preferences.append(pre)
         wow.append(a)
+    flag = False
+    if(sorted([student.budget for student in students])[-1] < maximum):
+        #the algorithm requires maximum to be less than maximum budget
+        logger.warning("maximum greater than max budget")
+    logger.debug("%g is the greatest budget", sorted([student.budget for student in students])[-1])
+
     csp_mapping(wow, courses)
     J_hat = sorted(courses, key=cmp_to_key(Course.comperator))[-1]
     ##till here line 1 in algorithm 2
@@ -86,6 +112,8 @@ def algorithm2(price_vector:list[float], maximum:int, eps:float, csp_mapping:cal
         d_star = math.floor((J_hat.capacity-J_hat.max_capacity)/2) #3
         p_l = J_hat.price #4
         p_h = maximum #5
+        logger.debug("low : %g", p_l)
+        logger.debug("high : %g", p_h)
         while(True): #6
             J_hat.price = (p_l + p_h)/2 #7
             ##same here as above, for line 8 if section
@@ -105,61 +133,8 @@ def algorithm2(price_vector:list[float], maximum:int, eps:float, csp_mapping:cal
                 J_hat.price = p_h #14
                 J_hat.mark = True #14
                 break #part of 13
+        logger.debug("new price %g for course %s", J_hat.price, J_hat.name)
         J_hat = sorted(courses, key=cmp_to_key(Course.comperator))[-1] #15
         # if(J_hat.capacity <= J_hat.max_capacity or maximum-J_hat.price <= eps):
         #     return [c.price for c in courses]
     return [c.price for c in courses] #return at the end
-
-
-def test1(): 
-    a = Course(name='a', price=9, max_capacity=5)
-    b = Course(name='b', price=2, max_capacity=3)
-    c = Course(name='c', price=4.5, max_capacity=5)
-    s1 = Student(name='s1', budget=15, preferences=([c, b]))
-    s2 = Student(name='s2', budget=15, preferences=([b, c, a]))
-    s3 = Student(name='s3', budget=15, preferences=([b, a]))
-    s4 = Student(name='s4', budget=15, preferences=([a, c]))
-    s5 = Student(name='s5', budget=15, preferences=([a, b, c]))
-    courses = [a, b, c]
-    students = [s1, s2, s3, s4, s5]
-    
-    eps = 1
-    maximum = 10
-    price_vector = [9,2,4.5]
-    assert algorithm2(price_vector, maximum, eps, csp_mapping, students, courses) == [9,10,4.5]
-
-def test2(): 
-    a = Course(name='a', price=2.4, capacity=0, max_capacity=5)
-    b = Course(name='b', price=5, capacity=0, max_capacity=4)
-    c = Course(name='c', price=10.4, capacity=0, max_capacity=5)
-    courses = [a, b, c]
-
-
-    s1 = Student(name='s1', budget=18, year=1, courses=[], preferences=([c, b]))
-    s2 = Student(name='s2', budget=18, year=1, courses=[], preferences=([b, c, a]))
-    s3 = Student(name='s3', budget=18, year=1, courses=[], preferences=([b, a]))
-    s4 = Student(name='s4', budget=18, year=1, courses=[], preferences=([a, c]))
-    s5 = Student(name='s5', budget=18, year=1, courses=[], preferences=([a, b, c]))
-    s6 = Student(name='s6', budget=18, year=1, courses=[], preferences=([a, c, b]))
-    students = [s1, s2, s3, s4, s5, s6]
-    eps = 1
-    maximum = 10
-    price_vector = [2.4,5,10.4]
-    assert algorithm2(price_vector, maximum, eps, csp_mapping, students, courses) == [2.4,8.125,10.4]
-
-
-def test3():
-    a = Course(name='a', price=3, max_capacity=5)
-    b = Course(name='b', price=4, max_capacity=3)
-    c = Course(name='c', price=3, max_capacity=5)
-    courses = [a, b, c]
-    s1 = Student(name='s1', budget=10, preferences=([c, b]))
-    s2 = Student(name='s2', budget=10, preferences=([b, c, a]))
-    s3 = Student(name='s3', budget=10, preferences=([b, a]))
-    s4 = Student(name='s4', budget=10, preferences=([a, b, c]))
-    s5 = Student(name='s5', budget=10, preferences=([a, b, c]))
-    students = [s1, s2, s3, s4, s5]
-    eps = 0.1
-    maximum = 6
-    price_vector = [3,4,3]
-    assert(algorithm2(price_vector, maximum, eps, csp_mapping, students, courses)==[3,6,3])
