+++ my-project-env/bin/jupyter-migrate
@@ -1,0 +1,8 @@
+#!/home/mrberlingo/Documents/Uni/fairpy/my-project-env/bin/python
+# -*- coding: utf-8 -*-
+import re
+import sys
+from jupyter_core.migrate import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
